+++ examples/image_streamer/deployment_plan.rb
@@ -1,4 +1,4 @@
-# (c) Copyright 2017 Hewlett Packard Enterprise Development LP
+# (c) Copyright 2020 Hewlett Packard Enterprise Development LP
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -9,6 +9,9 @@
 # CONDITIONS OF ANY KIND, either express or implied. See the License for the
 # specific language governing permissions and limitations under the License.
 
+# Defaults API version to 300
+node.default['oneview']['api_version'] = 300
+
 oneview_client = {
   url: ENV['ONEVIEWSDK_URL'],
   user: ENV['ONEVIEWSDK_USER'],
@@ -17,25 +20,24 @@ oneview_client = {
 
 i3s_client = {
   url: ENV['I3S_URL'],
-  oneview_client: oneview_client
+  oneview_client: oneview_client,
+  api_version: 1200
 }
 
 # Create or update the Deployment Plan named 'DeploymentPlan1'
 image_streamer_deployment_plan 'DeploymentPlan1' do
   client i3s_client
-  api_version 300
   data(
     description: 'AnyDescription',
     hpProvided: false
   )
-  os_build_plan 'ChefBP01'
-  golden_image 'ChefGI01'
+  os_build_plan 'Test-OSBuildPlan'
+  golden_image 'RHEl-capture'
 end
 
 # Update the Deployment Plan named 'DeploymentPlan1' with a new description
 image_streamer_deployment_plan 'DeploymentPlan1' do
   client i3s_client
-  api_version 300
   data(
     description: 'Chef created and updated Deployment Plan - 1'
   )
@@ -44,26 +46,23 @@ end
 # Create the Deployment Plan named 'DeploymentPlan2' only if does not exist
 image_streamer_deployment_plan 'DeploymentPlan2' do
   client i3s_client
-  api_version 300
   data(
     description: 'example of create_if_missing action',
     hpProvided: false
   )
-  os_build_plan 'ChefBP01'
-  golden_image 'ChefGI01'
+  os_build_plan 'RHEL-personalize-and-configure-NICs-LVM-BP-2018-09-11'
+  golden_image 'RHEl-capture'
   action :create_if_missing
 end
 
 # Delete the Deployment Plan named 'DeploymentPlan1' if it exists
 image_streamer_deployment_plan 'DeploymentPlan1' do
   client i3s_client
-  api_version 300
   action :delete
 end
 
 # Delete the Deployment Plan named 'DeploymentPlan2' if it exists
 image_streamer_deployment_plan 'DeploymentPlan2' do
   client i3s_client
-  api_version 300
   action :delete
 end