+++ spec/fixtures/cookbooks/image_streamer_test_api1200.rb/metadata.rb
@@ -1,0 +1,7 @@
+name             'image_streamer_test_api1200'
+maintainer       'none'
+license          'All rights reserved'
+description      'A test cookbook for the oneview cookbook ImageStreamer::API1200: module'
+version          '0.1.0'
+
+depends          'oneview'