+++ spec/fixtures/cookbooks/image_streamer_test_api800/metadata.rb
@@ -1,0 +1,7 @@
+name             'image_streamer_test_api800'
+maintainer       'none'
+license          'All rights reserved'
+description      'A test cookbook for the oneview cookbook ImageStreamer::API800: module'
+version          '0.1.0'
+
+depends          'oneview'