+++ spec/spec_helper.rb
@@ -109,4 +109,16 @@ RSpec.shared_context 'chef context', a: :b do
   let(:i3s_client600) do
     OneviewSDK::ImageStreamer::Client.new(url: 'https://i3s.example.com', token: 'token123', api_version: 600)
   end
+
+  let(:i3s_client800) do
+    OneviewSDK::ImageStreamer::Client.new(url: 'https://i3s.example.com', token: 'token123', api_version: 800)
+  end
+
+  let(:i3s_client1000) do
+    OneviewSDK::ImageStreamer::Client.new(url: 'https://i3s.example.com', token: 'token123', api_version: 1000)
+  end
+
+  let(:i3s_client1200) do
+    OneviewSDK::ImageStreamer::Client.new(url: 'https://i3s.example.com', token: 'token123', api_version: 1200)
+  end
 end